# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.UpdateFileAliasesRequest:
    properties:
      aliases:
        items:
          type: string
        type: array
      filePath:
        type: string
    required:
    - filePath
    type: object
  models.AliasModel:
    properties:
      alias:
        type: string
      filePath:
        type: string
      id:
        type: integer
      itemId:
        type: integer
      type:
        $ref: '#/definitions/schema.AliasType'
    required:
    - alias
    - id
    - type
    type: object
...
  schema.AliasType:
    enum:
    - epic
    - story
    - task
    - question
    - problem
    - knowledge-node
    - knowledge-bit
    - definition
    - action
    - repetitive-task
    - state
    - file
    type: string
    x-enum-varnames:
    - AliasTypeEpic
    - AliasTypeStory
    - AliasTypeTask
    - AliasTypeQuestion
    - AliasTypeProblem
    - AliasTypeKnowledgeNode
    - AliasTypeKnowledgeBit
    - AliasTypeDefinition
    - AliasTypeAction
    - AliasTypeRepetitiveTask
    - AliasTypeState
    - AliasTypeFile
...
  services.FileInfo:
    properties:
      isDir:
        type: boolean
      path:
        type: string
      size:
        type: integer
    type: object
  services.FilesConfig:
    properties:
      baseDir:
        type: string
      encrypted:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Task management dashboard API server.
    **Authentication:** Use `POST /users/login`, copy `accessToken` from the response, then click **Authorize** and paste it as the `access_token` cookie value. Alternatively, login from Swagger on the same host and cookies are sent automatically.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Dashboard API
  version: "1.0"
paths:
...
  /aliases/file:
    put:
      consumes:
      - application/json
      description: 'Replaces file aliases: removes absent ones and adds new ones.
        filePath is the logical relative path under FILES_DIR.'
      parameters:
      - description: Aliases sync request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateFileAliasesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync aliases for file
      tags:
      - aliases
  /aliases/file/{filepath}:
    get:
      consumes:
      - application/json
      description: Returns all aliases linked to the given logical relative file path
        under FILES_DIR
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List aliases for file
      tags:
      - aliases
...
  /files:
    get:
      description: Lists all files and directories under the configured files directory
        (logical relative paths; when FILES_ENCRYPTED=true the trailing .bin suffix
        is stripped)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.FileInfo'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List files
      tags:
      - files
  /files/config:
    get:
      description: Returns whether files are stored encrypted and the configured base
        directory
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.FilesConfig'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get files storage config
      tags:
      - files
...
  /files/content/{filepath}:
    delete:
      description: Deletes a file by logical relative path (when FILES_ENCRYPTED=true
        deletes path.bin)
      parameters:
      - description: Relative file path
        example: 1.txt
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a file
      tags:
      - files
    get:
      description: Returns file contents for the given logical relative path (e.g.
        1.txt or sub/dir/file.txt). When FILES_ENCRYPTED=true reads path.bin on disk
        and returns opaque ciphertext (no server-side decrypt).
      parameters:
      - description: Relative file path
        example: 1.txt
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get file by relative path
      tags:
      - files
    put:
      consumes:
      - application/octet-stream
      description: Writes request body to the given logical relative path. When FILES_ENCRYPTED=true
        stores as path.bin; client should send already-encrypted bytes.
      parameters:
      - description: Relative file path
        example: 1.txt
        in: path
        name: filepath
        required: true
        type: string
      - description: Raw file bytes
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload or replace a file
      tags:
      - files
  /files/parents-path/{filepath}:
    get:
      description: Resolves the owning container from a logical relative file path
        and returns its parent container descriptions (leaf→root), same shape as POST
        /parents-path
      parameters:
      - description: Relative file path
        example: tasks/12_foo/note.md
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parents path for a file
      tags:
      - files
...
schemes:
- http
- https
security:
- AccessTokenCookie: []
securityDefinitions:
  AccessTokenCookie:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"